int to float64 in stock and reposition

+++ cmd/adapters/entrypoints/handlers/material/entities.go
@@ -5,8 +5,8 @@ type MaterialResponse struct {
 	Name            string  `json:"name"`
 	Description     string  `json:"description"`
 	Price           float64 `json:"price"`
-	Stock           int     `json:"stock"`
-	RepositionPoint int     `json:"repositionPoint"`
+	Stock           float64 `json:"stock"`
+	RepositionPoint float64 `json:"repositionPoint"`
 	MaterialType    string  `json:"materialType"`
 	MaterialTypeId  uint    `json:"materialTypeId"`
 }
@@ -15,7 +15,7 @@ type MaterialRequest struct {
 	Name            string  `json:"name"`
 	Description     string  `json:"description"`
 	Price           float64 `json:"price"`
-	Stock           int     `json:"stock"`
-	RepositionPoint int     `json:"repositionPoint"`
+	Stock           float64 `json:"stock"`
+	RepositionPoint float64 `json:"repositionPoint"`
 	MaterialTypeId  uint    `json:"materialTypeId"`
 }
